+++ src/main/java/flow/extensionblocker/application/BlockerService.java
@@ -3,10 +3,10 @@ package flow.extensionblocker.application;
 import flow.extensionblocker.application.dto.BlockerResponse;
 import flow.extensionblocker.application.dto.CreateBlockerRequest;
 import flow.extensionblocker.application.dto.CreateBlockerResponse;
-import flow.extensionblocker.common.global.exception.BlockerAlreadyDeletedException;
-import flow.extensionblocker.common.global.exception.BlockerAlreadyExistsException;
-import flow.extensionblocker.common.global.exception.BlockerLimitExceededException;
-import flow.extensionblocker.common.global.exception.BlockerNotFoundException;
+import flow.extensionblocker.common.global.exception.blocker.BlockerAlreadyDeletedException;
+import flow.extensionblocker.common.global.exception.blocker.BlockerAlreadyExistsException;
+import flow.extensionblocker.common.global.exception.blocker.BlockerLimitExceededException;
+import flow.extensionblocker.common.global.exception.blocker.BlockerNotFoundException;
 import flow.extensionblocker.domain.Blocker;
 import flow.extensionblocker.domain.BlockerRepository;
 import jakarta.transaction.Transactional;
