+++ src/test/java/flow/extensionblocker/application/BlockerServiceTest.java
@@ -6,8 +6,8 @@ import static org.mockito.BDDMockito.any;
 import static org.mockito.BDDMockito.given;
 
 import flow.extensionblocker.application.dto.CreateBlockerRequest;
-import flow.extensionblocker.common.global.exception.BlockerAlreadyExistsException;
-import flow.extensionblocker.common.global.exception.BlockerLimitExceededException;
+import flow.extensionblocker.common.global.exception.blocker.BlockerAlreadyExistsException;
+import flow.extensionblocker.common.global.exception.blocker.BlockerLimitExceededException;
 import flow.extensionblocker.domain.Blocker;
 import flow.extensionblocker.domain.BlockerRepository;
 import flow.extensionblocker.domain.Type;
